File extraction and renaming
* We're not using minimist so rename minimist to argparse
* Extract files instead of having one big file that contain all the
  logic
