+++ src/index.ts
@@ -9,4 +9,8 @@
 import * as t from 'io-ts';
 export { t };
 export { unimplemented } from './utils';
+export { subcommands } from './subcommands';
+export { fromStr } from './fromStr';
+export { binaryParser } from './binaryParser';
 export * from './command';
+export { parse } from './parse';
