+++ src/utils.ts
@@ -1,4 +1,6 @@
+import chalk from 'chalk';
 import stripAnsi from 'strip-ansi';
+import { ParseItem } from './argparse';
 
 /**
  * Throws an error with an `unimplemented` message
@@ -19,4 +21,40 @@ export function padNoAnsi(str: string, length: number, place: 'end' | 'start') {
   } else {
     return pad + str;
   }
+}
+
+const colorCycle = [
+  chalk.green,
+  chalk.blue,
+  chalk.magenta,
+  chalk.cyan,
+  chalk.white,
+];
+
+export function generateColorCycle(): () => chalk.Chalk {
+  let i = 0;
+  return () => colorCycle[i++ % colorCycle.length];
+}
+
+export function contextToString(ctx: ParseItem[]): string {
+  const getColor = generateColorCycle();
+
+  let parts: string[] = [];
+  for (const item of ctx) {
+    switch (item.type) {
+      case 'positional': {
+        parts.push(item.input);
+        break;
+      }
+      case 'namedArgument': {
+        parts.push(`${item.inputKey} ${item.inputValue}`);
+        break;
+      }
+      case 'forcePositional': {
+        parts.push('--');
+        break;
+      }
+    }
+  }
+  return parts.map(x => getColor()(x)).join(' ');
 }
